Reorganize source into many smaller modules
