+++ src/CurrencyFormatting/templates.jl
